I changed the gitignore to ignore lock files.
Gitignore now ignores lock files. I added
.~lock.*#
to the .gitignore file. I have not yet tested this .gitignore to see if
it works with lock files. I do not know there are no issues. I will have
to go back and make changes to all the .gitignore based on the template
if the code does not work with lock files.